Move Linkage and Visibility to codegen_fn_attrs.
