+++ compiler/rustc_codegen_cranelift/src/linkage.rs
@@ -1,4 +1,5 @@
-use rustc_middle::mir::mono::{Linkage as RLinkage, MonoItem, Visibility};
+use rustc_middle::middle::codegen_fn_attrs::{Linkage as RLinkage, Visibility};
+use rustc_middle::mir::mono::MonoItem;
 
 use crate::prelude::*;
